% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocleHeatmap.R
\docType{data}
\name{BEAM_res}
\alias{BEAM_res}
\title{This is a test data for this package
test data describtion}
\format{
An object of class \code{data.frame} with 47192 rows and 8 columns.
}
\usage{
BEAM_res
}
\description{
This is a test data for this package
test data describtion
}
\author{
JunZhang
}
\keyword{datasets}
