% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-packages.R
\name{check_dependency}
\alias{check_dependency}
\title{Check and Install Required Packages}
\usage{
check_dependency()
}
\value{
NULL This function does not return a value. It installs the packages as needed.
}
\description{
This function checks if the required Bioconductor and CRAN packages are installed.
If any of the required packages are not installed, it installs them automatically.
This includes Bioconductor packages such as \code{clusterProfiler}, \code{SummarizedExperiment},
\code{ComplexHeatmap}, \code{SingleCellExperiment}, \code{TCseq}, \code{monocle}, and \code{Biobase}.
It also checks for CRAN packages such as \code{Seurat}, \code{WGCNA}, \code{igraph},
\code{pheatmap}, \code{circlize}, \code{e1071}, and \code{colorRamps}.
}
