% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterData.R
\name{clusterData}
\alias{clusterData}
\title{using clusterData to cluster genes}
\usage{
clusterData(
  exp = NULL,
  scaleData = TRUE,
  cluster.method = c("mfuzz", "TCseq", "kmeans", "wgcna"),
  TCseq_params_list = list(),
  object = NULL,
  min.std = 0,
  cluster.num = NULL,
  subcluster = NULL,
  seed = 5201314
)
}
\arguments{
\item{exp}{the gene normalized expression data(fpkm/rpkm/tpm), default NULL.}

\item{scaleData}{whether do Z-score for expression data, default TRUE.}

\item{cluster.method}{which cluster method to choose, mfuzz, kmeans or wgcna, default "mfuzz".
details see package Mfuzz. If choose "kmeans", the row_km argument will be used to cluster genes
in ComplexHeatmap::Heatmap function.}

\item{TCseq_params_list}{List papameters passed by TCseq::timeclust function.}

\item{object}{the WGCNA::blockwiseModules returned object, default NULL.}

\item{min.std}{min std for filter genes, this argument is in mfuzz function, default 0.}

\item{cluster.num}{the number clusters, default NULL.}

\item{subcluster}{subset some clusters form all clusterd data, default NULL.}

\item{seed}{set a seed for cluster analysis in mfuzz or Heatmap function, default 5201314.}
}
\value{
clusterData return a list including wide-shape and long-shape clustered results.
}
\description{
using clusterData to cluster genes
}
\examples{

data("exps")
# mfuzz
cm <- clusterData(exp = exps,
                  cluster.method = "mfuzz",
                  cluster.num = 8)

# kmeans
ck <- clusterData(exp = exps,
                  cluster.method = "kmeans",
                  cluster.num = 8)

}
\author{
JunZhang
}
