% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DataSet}
\alias{DataSet}
\alias{exps}
\alias{BEAM_res}
\alias{sig_gene_names}
\alias{net}
\alias{termanno}
\alias{termanno2}
\alias{pbmc_subset}
\title{This is a test data for this package}
\format{
A data frame.

A data frame.

A data frame.

A data frame.

A data frame.

A data frame.

A seurat object.
}
\description{
This is a test data for this package

This is a test data for this package

This is a test data for this package

This is a test data for this package

This is a test data for this package

This is a test data for this package

This is a test data for this package
}
\keyword{datasets}
