% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterData.R
\docType{data}
\name{exps}
\alias{exps}
\title{This is a test data for this package
test data describtion}
\format{
An object of class \code{data.frame} with 3767 rows and 6 columns.
}
\usage{
exps
}
\description{
This is a test data for this package
test data describtion
}
\author{
Junjun Lao
}
\keyword{datasets}
