% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.std.R
\name{filter.std modified by Mfuzz filter.std}
\alias{filter.std modified by Mfuzz filter.std}
\alias{filter.std}
\title{using filter.std to filter low expression genes}
\usage{
filter.std(eset, min.std, visu = TRUE, verbose = TRUE)
}
\arguments{
\item{eset}{expression matrix, default NULL.}

\item{min.std}{min stand error, default 0.}

\item{visu}{whether plot, default FALSE.}

\item{verbose}{show filter information.}
}
\value{
matrix.
}
\description{
using filter.std to filter low expression genes
}
