% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.std.R
\name{filter.std}
\alias{filter.std}
\title{using filter.std to filter low expression genes}
\usage{
filter.std(eset = NULL, min.std = 0, visu = FALSE)
}
\arguments{
\item{eset}{expression matrix, default NULL.}

\item{min.std}{min stand error, default 0.}

\item{visu}{whether plot, default FALSE.}
}
\value{
matrix.
}
\description{
using filter.std to filter low expression genes
}
