% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusters.R
\name{getClusters}
\alias{getClusters}
\title{using getClusters to define a optimal cluster numbers}
\usage{
getClusters(exp = NULL)
}
\arguments{
\item{exp}{the gene normalized expression data(fpkm/rpkm/tpm), default NULL.}
}
\value{
a ggplot.
}
\description{
using getClusters to define a optimal cluster numbers
}
\examples{
\dontrun{
data("exps")
getClusters(exps)
}
}
\author{
JunZhang
}
