% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocleHeatmap.R
\docType{data}
\name{HSMM}
\alias{HSMM}
\title{This is a test data for this package
test data describtion}
\format{
An object of class \code{CellDataSet} with 47192 rows and 271 columns.
}
\usage{
HSMM
}
\description{
This is a test data for this package
test data describtion
}
\author{
JunZhang
}
\keyword{datasets}
