% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.enrichCluster.R
\docType{data}
\name{net}
\alias{net}
\title{This is a test data for this package
test data describtion}
\format{
An object of class \code{list} of length 10.
}
\usage{
net
}
\description{
This is a test data for this package
test data describtion
}
\author{
Junjun Lao
}
\keyword{datasets}
