% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocleHeatmap.R
\docType{data}
\name{sig_gene_names}
\alias{sig_gene_names}
\title{This is a test data for this package
test data describtion}
\format{
An object of class \code{character} of length 1331.
}
\usage{
sig_gene_names
}
\description{
This is a test data for this package
test data describtion
}
\author{
JunZhang
}
\keyword{datasets}
