% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visCluster.R
\docType{data}
\name{termanno2}
\alias{termanno2}
\title{This is a test data for this package
test data describtion}
\format{
An object of class \code{data.frame} with 24 rows and 3 columns.
}
\usage{
termanno2
}
\description{
This is a test data for this package
test data describtion
}
\author{
Junjun Lao
}
\keyword{datasets}
