% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocleHeatmap.R
\name{traverseTree}
\alias{traverseTree}
\title{traverseTree function}
\usage{
traverseTree(g, starting_cell, end_cells)
}
\arguments{
\item{g}{NULL}

\item{starting_cell}{NULL}

\item{end_cells}{NULL}
}
\description{
traverseTree function
}
