% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichCluster.R
\name{enrichCluster}
\alias{enrichCluster}
\title{using enrichCluster to do GO/KEGG enrichment analysis for multiple cluster genes}
\arguments{
\item{object}{clusterData object, default NULL.}

\item{type}{which type to choose for enrichment "BP","MF","CC","KEGG" or "ownSet".}

\item{TERM2GENE}{user input annotation of TERM TO GENE mapping, a data.frame
of 2 column with term and gene, default NULL.}

\item{TERM2NAME}{user input of TERM TO NAME mapping, a data.frame of 2 column
with term and name, default NULL.}

\item{OrgDb}{the annotation data for enrichment, default NULL.}

\item{id.trans}{whether perform the ID transformation, default TRUE.}

\item{fromType}{the input ID type, default "SYMBOL".}

\item{toType}{the ID type for "bitr" function to transform, default c("ENTREZID").}

\item{readable}{whether make the enrichmemnt results ID readble, default TRUE.}

\item{organism}{the organism name for "KEGG" enrichment,mouse("mmu"), human("hsa"), default NULL.}

\item{pvalueCutoff}{pvalueCutoff for enrichment, default 0.05.}

\item{topn}{the top enrichment results to extract, length one or same with cluster numbers, default 5.}

\item{seed}{the enrichment seed, default 5201314.}

\item{add.gene}{whether return genes for output, default FALSE.}
}
\value{
a data.frame.
}
\description{
using enrichCluster to do GO/KEGG enrichment analysis for multiple cluster genes
}
\examples{
\dontrun{
library(org.Mm.eg.db)

data("exps")

# mfuzz
ck <- clusterData(exp = exps,
                  cluster.method = "kmeans",
                  cluster.num = 8)

# enrich for clusters
enrich <- enrichCluster(object = ck,
                        OrgDb = org.Mm.eg.db,
                        type = "BP",
                        topn = 5)
}
}
\author{
JunZhang
}
