% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visCluster.R
\name{visCluster}
\alias{visCluster}
\title{using visCluster to visualize cluster results from clusterData output}
\arguments{
\item{object}{clusterData object, default NULL.}

\item{plot.type}{the plot type to choose which incuding "line","heatmap" and "both".}

\item{ms.col}{membership line color form Mfuzz cluster method results,
default c('#0099CC','grey90','#CC3333').}

\item{line.size}{line size for line plot, default 0.1.}

\item{line.col}{line color for line plot, default "grey90".}

\item{add.mline}{whether add median line on plot, default TRUE.}

\item{mline.size}{median line size, default 2.}

\item{mline.col}{median line color, default "#CC3333".}

\item{ncol}{the columns for facet plot with line plot, default 4.}

\item{ctAnno.col}{the heatmap cluster annotation bar colors, default NULL.}

\item{set.md}{the represent line method on heatmap-line plot(mean/median), default "median".}

\item{textbox.pos}{the relative position of text in left-line plot, default c(0.5,0.8).}

\item{textbox.size}{the text size of the text in left-line plot, default 8.}

\item{panel.arg}{the settings for the left-line panel which are
panel size,gap,width,fill and col, default c(2,0.25,4,"grey90",NA).}

\item{annoTerm.data}{the GO term annotation for the clusters, default NULL.}

\item{termAnno.arg}{the settings for GO term panel annotations which are fill and col,
default c("grey95","grey50").}

\item{...}{othe aruguments passed by Heatmap fuction.}
}
\value{
a ggplot2 or Heatmap object.
}
\description{
using visCluster to visualize cluster results from clusterData output
}
\examples{
\dontrun{
data("termanno")
data("exps")

# mfuzz
cm <- clusterData(exp = exps,
                  cluster.method = "mfuzz",
                  cluster.num = 8)

# plot
visCluster(object = cm,
           plot.type = "line")
}
}
\author{
JunZhang
}
