% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visCluster.R
\name{visCluster}
\alias{visCluster}
\title{using visCluster to visualize cluster results from clusterData output}
\arguments{
\item{object}{clusterData object, default NULL.}

\item{ht.col.list}{list of heatmap col_range and col_color, default
list(col_range = c(-2, 0, 2),col_color = c("#08519C", "white", "#A50F15")).}

\item{border}{whether add border for heatmap, default TRUE.}

\item{plot.type}{the plot type to choose which incuding "line","heatmap" and "both".}

\item{ms.col}{membership line color form Mfuzz cluster method results,
default c('#0099CC','grey90','#CC3333').}

\item{line.size}{line size for line plot, default 0.1.}

\item{line.col}{line color for line plot, default "grey90".}

\item{add.mline}{whether add median line on plot, default TRUE.}

\item{mline.size}{median line size, default 2.}

\item{mline.col}{median line color, default "#CC3333".}

\item{ncol}{the columns for facet plot with line plot, default 4.}

\item{ctAnno.col}{the heatmap cluster annotation bar colors, default NULL.}

\item{set.md}{the represent line method on heatmap-line plot(mean/median), default "median".}

\item{textbox.pos}{the relative position of text in left-line plot, default c(0.5,0.8).}

\item{textbox.size}{the text size of the text in left-line plot, default 8.}

\item{panel.arg}{the settings for the left-line panel which are
panel size,gap,width,fill and col, default c(2,0.25,4,"grey90",NA).}

\item{annoTerm.data}{the GO term annotation for the clusters, default NULL.}

\item{annoTerm.mside}{the wider GO term annotation box side, default "right".}

\item{termAnno.arg}{the settings for GO term panel annotations which are fill and col,
default c("grey95","grey50").}

\item{add.box}{whether add boxplot, default FALSE.}

\item{boxcol}{the box fill colors, default NULL.}

\item{box.arg}{this is related to boxplot width and border color, default c(0.1,"grey50").}

\item{add.point}{whether add point, default FALSE.}

\item{point.arg}{this is related to point shape,fill,color and size, default c(19,"orange","orange",1).}

\item{add.line}{whether add line, default TRUE.}

\item{line.side}{the line annotation side, default "right".}

\item{markGenes}{the gene names to be added on plot, default NULL.}

\item{markGenes.side}{the gene label side, default "right".}

\item{genes.gp}{gene labels graphics settings, default c('italic',10,NA).}

\item{go.col}{the GO term text colors, default NULL.}

\item{go.size}{the GO term text size(numeric or "pval"), default NULL.}

\item{mulGroup}{to draw multiple lines annotation, supply the groups numbers with vector, default NULL.}

\item{lgd.label}{the lines annotation legend labels, default NULL.}

\item{show_row_names}{whether to show row names, default FALSE.}

\item{term.text.limit}{the GO term text size limit, default c(10,18).}

\item{subgroup.anno}{the sub-cluster for annotation, supply sub-cluster id, default NULL.}

\item{add.bar}{whether add bar plot for GO enrichment, default FALSE.}

\item{bar.width}{the GO enrichment bar width, default 8.}

\item{textbar.pos}{the barplot text relative position, default c(0.8,0.8).}

\item{annnoblock.text}{whether add cluster numbers on right block annotation, default TRUE.}

\item{annnoblock.gp}{right block annotation text color and size, default c("white",8).}

\item{add.sampleanno}{whether add column annotation, default TRUE.}

\item{sample.group}{the column sample groups, default NULL.}

\item{sample.col}{column annotation colors, default NULL.}

\item{sample.order}{the orders for column samples, default NULL.}

\item{HeatmapAnnotation}{the 'HeatmapAnnotation' object from 'ComplexHeatmap'
when you have multiple annotations, default NULL.}

\item{column.split}{how to split the columns when supply multiple column annotations, default NULL.}

\item{cluster.order}{the row cluster orders for user's own defination, default NULL.}

\item{sample.cell.order}{the celltype order when input is scRNA data and "showAverage = FALSE"
for prepareDataFromscRNA.}

\item{annoKegg.data}{the KEGG term annotation for the clusters, default NULL.}

\item{annoKegg.mside}{the wider KEGG term annotation box side, default "right".}

\item{keggAnno.arg}{the settings for KEGG term panel annotations which are fill and col,
default c("grey95","grey50").}

\item{add.kegg.bar}{whether add bar plot for KEGG enrichment, default FALSE.}

\item{kegg.col}{the KEGG term text colors, default NULL.}

\item{kegg.size}{the KEGG term text size(numeric or "pval"), default NULL.}

\item{by.go}{the GO term text box style("anno_link" or "anno_block"), default "anno_link".}

\item{by.kegg}{the KEGG term text box style("anno_link" or "anno_block"), default "anno_link".}

\item{word_wrap}{whether wrap the text, default TRUE.}

\item{add_new_line}{whether add new line when text is long, default TRUE.}

\item{...}{othe aruguments passed by Heatmap fuction.}
}
\value{
a ggplot2 or Heatmap object.
}
\description{
using visCluster to visualize cluster results from clusterData output
}
\examples{
\dontrun{
data("termanno")
data("exps")

# mfuzz
cm <- clusterData(exp = exps,
                  cluster.method = "mfuzz",
                  cluster.num = 8)

# plot
visCluster(object = cm,
           plot.type = "line")
}
}
\author{
JunZhang
}
